+++ backend/app/Filament/Resources/SiteSettings/SiteSettingResource.php
@@ -20,7 +20,7 @@ class SiteSettingResource extends Resource
 
     protected static string|BackedEnum|null $navigationIcon = Heroicon::OutlinedRectangleStack;
 
-    protected static ?string $recordTitleAttribute = 'site_name';
+    protected static ?string $recordTitleAttribute = 'seo_title';
 
     public static function form(Schema $schema): Schema
     {